@@ -17,9 +17,12 @@
   },
   "homepage": "https://github.com/jaredpattison/dsa#readme",
   "dependencies": {
+    "eslint": "^5.15.3"
+  },
+  "devDependencies": {
     "@types/jest": "^24.0.11",
-    "eslint": "^5.15.3",
     "jest": "^24.5.0",
-    "typescript": "^3.3.4000"
+    "ts-jest": "^24.0.1",
+    "typescript": "^3.4.1"
   }
 }
